fix(feeRate): 직접 입력 - 소수점 가능

+++ lib/model/send/fee_info.dart
@@ -16,7 +16,7 @@ class FeeInfoWithLevel extends FeeInfo {
 class FeeInfo {
   int? estimatedFee;
   int? fiatValue;
-  int? satsPerVb;
+  double? satsPerVb;
   bool failedEstimation;
   bool isEstimating;
